feat: move rooms dashboard from / to /rooms
Root "/" is generic; give the dashboard its own path. "/" now
redirects to "/rooms" so bookmarks and existing links keep working.
Updated all internal navigation (login, invite accept/decline,
delete/exit room, navbar) to target /rooms instead of /.
