+++ src/app/rooms/_components/RoomsPage.jsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
-import { getUserRooms } from '../actions';
+import { getUserRooms } from '../../actions';
 import Box from '@mui/joy/Box';
 import CircularProgress from '@mui/joy/CircularProgress';
 import Button from '@mui/joy/Button';
